chore(frontend): move searchfield component
