+++ apps/frontend/src/components/common/SearchField.tsx
